RST-1072 Add missing repositories from hotdog (#17)
* Support custom distributions

* Add missing packages from hotdog to tailor-distro

* Only build dev flavour

* Add a bunch of entries

* No more missing deps!

* Use clean flag

* Catch github exceptions

* Fix pulling down pinned repos

* Fix things

* Fixup pins for fastrtps

* Pin all non-locus packages

* Change ROS2 upstream

* Switch repo for ceres

* Remove metapackages and rqt_robot_plugins

* Use gazebo9

* Add gazebo sources

* Fork open_karto to release c++11 fix

* Fix sparse_bundle_adjustment

* Add opencv binary

* Remove variant

* Use ros_comm_msgs fork

* Last straw with opencv3

* Include laser_proc patch; include shape in xenial opencv

* Disable bionic for now

* Bump opencv3

* Disable ccache

* Screw building opencv

* Try new gator version

* Set fext-numeric-literals globally

* Re-fork gazebo_ros_pkgs

* Fix slam_karto version

* Fix version of slam_karto

* Add a gazebo test bundle

* Use cpp14-compliant branches

* Update with s_b_a branch

* Re-enable dev flavour

* Remove karto recipe

* Add missing ros_control packages

* Update kdl_parser

* Fixup compare script

* Unpin realtime_tools

* Implement package whitelisting

* Repin realtime tools

* Remove extra imports

* Remove navigation metapackage

* Fix imu_tools

* Limit package builds

* White list grid_map packages

* Only build every 3 hours

* Restore flavours

* Drop ceres package

* Cleanup

+++ tailor_distro/pull_distro_repositories.py
@@ -3,28 +3,73 @@ import argparse
 import click
 import pathlib
 import rosdistro
-import subprocess
 import sys
-import yaml
+import github
+import tarfile
+import glob
+import shutil
 
+from catkin_pkg.package import parse_package
+from concurrent.futures import ThreadPoolExecutor
 from jinja2 import Environment, BaseLoader
 from shutil import rmtree
 from typing import Any, Mapping
-from urllib.parse import urlsplit, urlunsplit
+from urllib.parse import urlsplit
+from urllib import request
 
 from . import YamlLoadAction
 
 
-def process_url(url: str, organization: str, github_key: str = None) -> str:
-    """Process URL to prepend an authentication key."""
-    pieces = urlsplit(url)
-    if github_key is not None and pieces.netloc == 'github.com' and pieces.path.startswith('/' + organization):
-        pieces = pieces._replace(netloc=github_key + '@' + pieces.netloc)
+def pull_repository(repo_name, distro_data, repo_dir, github_client, upstream_name):
+    click.echo(f'Pulling repository {repo_name} ...', err=True)
+    repo_dir.mkdir(parents=True, exist_ok=True)
 
-    return urlunsplit(pieces)
+    try:
+        url = distro_data.release_repository.url
+    except AttributeError:
+        url = distro_data.source_repository.url
+
+    try:
+        version_template = distro_data.release_repository.tags['release']
+        context = {
+            'package': repo_name,
+            'upstream': upstream_name,
+            'version': distro_data.release_repository.version
+        }
+        version = Environment(loader=BaseLoader()).from_string(version_template).render(**context)
+    except (AttributeError, KeyError):
+        version = distro_data.source_repository.version
+
+    try:
+        package_names = distro_data.release_repository.package_names
+        if package_names != [repo_name]:
+            package_whitelist = package_names
+    except AttributeError:
+        package_whitelist = None
+
+    # TODO(pbovbel) Abstract interface away for github/bitbucket/gitlab
+    gh_repo_name = urlsplit(url).path[len('/'):-len('.git')]
+    gh_repo = github_client.get_repo(gh_repo_name, lazy=False)
+    archive_url = gh_repo.get_archive_link('tarball', version)
+
+    archive_file = repo_dir / f'{repo_name}.tar.gz'
+    with open(archive_file, 'wb') as tarball:
+        tarball.write(request.urlopen(archive_url).read())
+    with tarfile.open(archive_file) as tar:
+        tar.extractall(path=repo_dir)
+
+    # Remove all except whitelisted packages
+    if package_whitelist:
+        found_packages = glob.glob(str(repo_dir / '**/package.xml'), recursive=True)
+        for package_xml_path in found_packages:
+            package = parse_package(package_xml_path)
+            if package.name not in package_whitelist:
+                click.echo(f'Removing {package.name}, not in whitelist', err=True)
+                shutil.rmtree(pathlib.Path(package_xml_path).parent.resolve())
 
 
-def pull_distro_repositories(src_dir: pathlib.Path, recipes: Mapping[str, Any], github_key: str = None) -> None:
+def pull_distro_repositories(
+        src_dir: pathlib.Path, recipes: Mapping[str, Any], clean: bool, github_key: str = None) -> None:
     """Pull all the packages in all ROS distributions to disk
     :param src_dir: Directory where sources should be pulled.
     :param recipes: Recipe configuration defining distributions.
@@ -32,62 +77,26 @@ def pull_distro_repositories(src_dir: pathlib.Path, recipes: Mapping[str, Any], 
     """
     index = rosdistro.get_index(rosdistro.get_index_url())
 
+    github_client = github.Github(github_key)
+
     common_options = recipes['common']
-    organization = common_options['organization']
 
-    for distro_name in common_options['distributions'].keys():
+    with ThreadPoolExecutor() as executor:
 
-        click.echo(f"Pulling distribution {distro_name} ...", err=True)
+        for distro_name, distro_options in common_options['distributions'].items():
 
-        distro = rosdistro.get_distribution(index, distro_name)
-        target_dir = src_dir / distro_name
+            distro = rosdistro.get_distribution(index, distro_name)
+            target_dir = src_dir / distro_name
 
-        repositories = {}
-        for repo, data in distro.repositories.items():
+            if clean and target_dir.exists():
+                rmtree(str(target_dir))
 
-            context = {
-                'package': repo,
-                'upstream': common_options['distributions'][distro_name]['upstream'],
-            }
+            target_dir.mkdir(parents=True, exist_ok=not clean)
 
-            if data.release_repository:
-                url = data.release_repository.url
-                version = data.release_repository.tags['release']
-                type = data.release_repository.type
-                context['version'] = data.release_repository.version
-
-                # TODO(pbovbel) implement package whitelist/blacklist via release_repository.packages
-            elif data.source_repository:
-                url = data.source_repository.url
-                version = data.source_repository.version
-                type = data.source_repository.type
-            else:
-                click.echo(click.style(f"No source or release entry for {repo}", fg='yellow'), err=True)
-                continue
-
-            repositories[repo] = {
-                'type': type,
-                'url': process_url(url, organization, github_key),
-                'version': Environment(loader=BaseLoader()).from_string(version).render(**context)
-            }
-
-        if target_dir.exists():
-            rmtree(str(target_dir))
-
-        target_dir.mkdir(parents=True)
-
-        repositories_file = src_dir / (distro_name + '.repos')
-        repositories_file.write_text(yaml.dump({'repositories': repositories}))
-
-        vcs_command = [
-            "vcs", "import", str(target_dir),
-            "--input", str(repositories_file),
-            "--retry", str(3),
-            # "--recursive",
-            "--shallow",
-        ]
-        click.echo(' '.join(vcs_command), err=True)
-        subprocess.run(vcs_command, check=True)
+            for repo_name, distro_data in distro.repositories.items():
+                repo_dir = target_dir / repo_name
+                executor.submit(pull_repository, repo_name, distro_data, repo_dir, github_client,
+                                upstream_name=distro_options['upstream']['name'])
 
 
 def main():
@@ -95,6 +104,7 @@ def main():
     parser.add_argument('--src-dir', type=pathlib.Path, required=True)
     parser.add_argument('--recipes', action=YamlLoadAction, required=True)
     parser.add_argument('--github-key', type=str)
+    parser.add_argument('--clean', action='store_true')
     args = parser.parse_args()
 
     sys.exit(pull_distro_repositories(**vars(args)))
